packages/mc_lightning/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=0.10.0=py38h32f6830_0
aiohttp=3.6.2=py38h516909a_0
async-timeout=3.0.1=py_1000
attrs=20.2.0=pyh9f0ad1d_0
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py38h1e0a361_1000
c-ares=1.16.1=h516909a_3
ca-certificates=2020.6.20=hecda079_0
cachetools=4.1.1=py_0
certifi=2020.6.20=py38h32f6830_0
cffi=1.14.3=py38h5bae8af_0
chardet=3.0.4=py38h32f6830_1007
click=7.1.2=pyh9f0ad1d_0
cryptography=3.1.1=py38h766eaa4_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
freetype=2.10.2=he06d7ca_0
future=0.18.2=py38h32f6830_1
google-auth=1.22.0=py_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.31.0=py38h2c89da0_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=2.0.0=py38h32f6830_0
joblib=0.17.0=py_0
jpeg=9d=h516909a_0
kiwisolver=1.2.0=pypi_0
lcms2=2.11=hbd6801e_0
ld_impl_linux-64=2.35=h769bd43_9
libblas=3.8.0=16_mkl
libcblas=3.8.0=16_mkl
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=7.5.0=hae1eefd_17
libgfortran4=7.5.0=hae1eefd_17
liblapack=3.8.0=16_mkl
libpng=1.6.37=hed695b0_2
libprotobuf=3.13.0=h8b12597_0
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtiff=4.1.0=hc7e4089_6
libwebp-base=1.1.0=h516909a_3
llvm-openmp=10.0.1=hc9558a2_0
lz4-c=1.9.2=he1b5a44_3
markdown=3.2.2=py_0
matplotlib=3.3.2=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py38h1e0a361_1
mkl_fft=1.2.0=py38hab2c0dc_1
mkl_random=1.2.0=py38hc5bc63f_1
multidict=4.7.5=py38h1e0a361_1
ncurses=6.2=he1b5a44_1
ninja=1.10.1=hc9558a2_1
numpy=1.19.1=py38hbc911f0_0
numpy-base=1.19.1=py38hfa32c7d_0
oauthlib=3.0.1=py_0
olefile=0.46=py_0
openssl=1.1.1h=h516909a_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.1.2=pypi_0
pillow=7.2.0=py38h9776b28_1
pip=20.2.3=py_0
protobuf=3.13.0=py38h950e882_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py38h32f6830_1
python=3.8.5=h1103e12_9_cpython
python-dateutil=2.8.1=pypi_0
python_abi=3.8=1_cp38
pytorch=1.6.0=py3.8_cuda10.1.243_cudnn7.6.3_0
pytorch-lightning=0.9.0=py_0
pytz=2020.1=pypi_0
pyyaml=5.3.1=py38h1e0a361_0
readline=8.0=he28a2e2_2
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.6=pyh9f0ad1d_0
scikit-learn=0.23.2=py38hee58b96_0
scipy=1.5.2=py38h8c5af15_0
setuptools=49.6.0=py38h32f6830_1
six=1.15.0=pyh9f0ad1d_0
sqlite=3.33.0=h4cf870e_0
tensorboard=2.2.0=py_0
tensorboard-plugin-wit=1.6.0=pyh9f0ad1d_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hed695b0_0
torchvision=0.7.0=py38_cu101
tqdm=4.50.0=pyh9f0ad1d_0
urllib3=1.25.10=py_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.35.1=pyh9f0ad1d_0
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
yarl=1.4.2=py38h516909a_0
zipp=3.3.0=py_0
zlib=1.2.11=h516909a_1009
zstd=1.4.5=h6597ccf_2
